--- a/composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/recording/plot/spectrum/SpectrumPlotView.kt
+++ b/composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/recording/plot/spectrum/SpectrumPlotView.kt
@@ -2,6 +2,7 @@ package org.noiseplanet.noisecapture.ui.features.recording.plot.spectrum
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.EaseOutBack
+import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
@@ -14,13 +15,16 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.BiasAlignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
@@ -29,11 +33,10 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import org.koin.compose.viewmodel.koinViewModel
 import org.noiseplanet.noisecapture.util.toFrequencyString
 import kotlin.math.max
-import kotlin.math.min
 
 
 @Composable
-fun SpectrumPlotViewCompose(
+fun SpectrumPlotView(
     backgroundColor: Color = MaterialTheme.colorScheme.surface,
     modifier: Modifier = Modifier,
 ) {
@@ -44,9 +47,27 @@ fun SpectrumPlotViewCompose(
     val axisSettings by viewModel.axisSettingsFlow.collectAsStateWithLifecycle()
     val splData by viewModel.splDataFlow.collectAsStateWithLifecycle()
 
-    val gradientBrush = Brush.horizontalGradient(
-        *viewModel.spectrumColorRamp.toTypedArray()
-    )
+    // How wide is each frequency band bar relative to the width of the plot width
+    val rawSplBarWidths by derivedStateOf {
+        splData.mapValues { (_, spl) ->
+            (spl.raw / axisSettings.maximumX).toFloat()
+        }.values.reversed()
+    }
+
+    // Offset of each weighted spl value per frequency band (-1 is left aligned, 1 is right aligned)
+    val weightedSplBoxBiases by derivedStateOf {
+        splData.mapValues { (_, spl) ->
+            ((max(spl.weighted, spl.raw) / axisSettings.maximumX) * 2.0 - 1.0).toFloat()
+        }.values.reversed()
+    }
+
+    // Gradient brush to paint the plot background
+    val gradientBrush = remember {
+        Brush.horizontalGradient(
+            *viewModel.spectrumColorRamp.toTypedArray()
+        )
+    }
+
 
     // - Layout
 
@@ -58,15 +79,11 @@ fun SpectrumPlotViewCompose(
                 .background(gradientBrush)
         ) {
             // 2. Adjust bars size by clipping from the end of the screen (i.e. drawing the negative)
-            for (spl in splData.values.reversed()) {
-                val widthFraction =
-                    (min(spl.raw, axisSettings.maximumX) / axisSettings.maximumX).toFloat()
+            for (widthFraction in rawSplBarWidths) {
                 Box(
                     modifier = Modifier.weight(1f)
                         .background(backgroundColor)
-                        .animateContentSize(
-                            tween(durationMillis = 200, easing = EaseOutBack)
-                        )
+                        .animateContentSize(tween(easing = EaseOutBack, durationMillis = 400))
                         .fillMaxWidth(fraction = 1f - widthFraction)
                 )
             }
@@ -81,30 +98,23 @@ fun SpectrumPlotViewCompose(
 
         // 4. Draw dark rectangles to show weighted spl values
         Column(
-            verticalArrangement = Arrangement.spacedBy(2.dp),
+            verticalArrangement = Arrangement.spacedBy(0.dp),
             modifier = Modifier.fillMaxSize()
         ) {
-            for (spl in splData.values.reversed()) {
-                val widthFraction = (max(spl.weighted, spl.raw) / axisSettings.maximumX).toFloat()
-                val boxColor = MaterialTheme.colorScheme.onSurface
+            val boxColor = MaterialTheme.colorScheme.onSurface
+
+            weightedSplBoxBiases.forEach { bias ->
+                val animatedBias by animateFloatAsState(
+                    targetValue = bias,
+                    animationSpec = tween(easing = EaseOutBack, durationMillis = 400)
+                )
 
                 Box(
-                    modifier = Modifier.weight(1f)
-                        .background(Color.Transparent)
-                        .animateContentSize(
-                            tween(durationMillis = 200, easing = EaseOutBack)
-                        )
-                        .fillMaxWidth(fraction = widthFraction)
-                        .drawBehind {
-                            val boxWidth = 10.dp.toPx()
-                            drawRe
-                            drawRect(
-                                color = boxColor,
-                                start = Offset(size.width, 0f),
-                                end = Offset(size.width, size.height),
-                                strokeWidth = boxWidth
-                            )
-                        }
+                    modifier = Modifier.width(10.dp)
+                        .padding(bottom = 2.dp)
+                        .weight(1f)
+                        .background(boxColor)
+                        .align(BiasAlignment.Horizontal(animatedBias))
                 )
             }
         }
@@ -112,6 +122,9 @@ fun SpectrumPlotViewCompose(
 }
 
 
+/**
+ * Lays out a grid of given X and Y lines with given line color.
+ */
 @Composable
 private fun SpectrumPlotAxisGrid(
     xAxisTicks: Int,
@@ -121,7 +134,7 @@ private fun SpectrumPlotAxisGrid(
     val strokeWidth = 2.dp.toPx()
     val yOffsetStep = size.height / yAxisTicks
     var yOffset = yOffsetStep - strokeWidth / 2f
-    repeat(yAxisTicks - 1) {
+    repeat(yAxisTicks) {
         drawLine(
             color = lineColor,
             start = Offset(0f, yOffset),
@@ -146,6 +159,9 @@ private fun SpectrumPlotAxisGrid(
 }
 
 
+/**
+ * Lays out X and Y axes with given content.
+ */
 @Composable
 private fun SpectrumPlotContainer(
     axisSettings: SpectrumPlotViewModel.AxisSettings,