# Description

Rewrote `SpectrumPlotView` using composable instead of manual Canvas drawing. This allows to fix some caveats of the old method:

- Bars and boxes can smoothly animate between two states (instead of being limited to 8 fps (125ms)). See GIFs below for comparison
- Placement of axis labels is calculated automatically during first composition (less manual logic to handle on our side)

| Before | After |
| ------ | ----- |
| ![ScreenRecording2025-07-10at11 08 36-ezgif com-optimize](https://github.com/user-attachments/assets/703f37b0-1dbc-4925-a3c3-9bf000bd519e) | ![ScreenRecording2025-07-10at11 46 13-ezgif com-optimize](https://github.com/user-attachments/assets/209e7a89-7c05-4aa5-88a3-cf67d82f0e20) |

It was maybe not top priority but I wanted to try it out now that I have a bit more Compose knowledge 😁 

## Changes

- Axes are placed using `Row` and `Column`
- Plot area is painted once using the noise color ramp gradient
- Bars are drawn negatively (right aligned inverse value) on top of the background. Width is animated between two values.
- Axis grids are drawn on top of the bars to add vertical and horizontal dividers
- Weighted SPL boxes are drawn using `Modifier.align(BiasAlignment.Horizontal(bias))` for relative positionning

## Linked issues

- #128 

## Remaining TODOs

N/A

## Checklist

- [x] Code compiles correctly on all platforms
- [x] All pre-existing tests are passing
- [ ] If needed, new tests have been added
- [ ] Extended the README / documentation if necessary
- [x] Added code has been documented