@nicolas-f 2025-07-15
Nice ^^
some feedback:

- Ticks are not visible
- Position of graphic elements are not clamped in the drawing area:
<img width="1080" height="2400" alt="image" src="https://github.com/user-attachments/assets/b3ef70b5-a1b3-451c-9cb5-d0140bf696af" />