@tonelli-m 2025-07-15
Thanks for the feedback! 
I've clipped the plot to the container bounds, and added ticks to the x axis labels. I think it's a bit redundant for the y axis tho because every frequency band already has its label.